ref(core): Merge browser and server spanStreamingIntegration into one shared integration
The browser and server span streaming integrations only differed in their
eager-flush trigger, so collapse them into a single core integration
parameterized by a `flushOnSegmentEnd` option. The browser install site
(`browserSpanApi`) passes `{ flushOnSegmentEnd: true }`; the server side
(`ServerRuntimeClient`) keeps the same call and behavior.

The integration is now exported from `shared-exports` (a single shared symbol
across all entries) instead of being duplicated in `browser-exports` and
`server-exports`. Because both entries previously resolved to *different*
symbols, the name was an ambiguous star export that `index.ts` disambiguated
explicitly; now that it resolves to the same symbol, that explicit re-export
collided with the star exports and is removed.
