+++ packages/core/src/browser-exports.ts
@@ -13,8 +13,6 @@ export {
 
 export { startIdleSpan } from './tracing/idleSpan';
 
-export { spanStreamingIntegration } from './integrations/browserSpanStreaming';
-
 export { supportsNativeFetch } from './utils/supports';
 export type { XhrBreadcrumbData, XhrBreadcrumbHint } from './types/breadcrumb';
 export type { BrowserClientReplayOptions } from './types/browseroptions';
