+++ packages/core/src/integrations/spanStreaming.ts
@@ -9,7 +9,20 @@ import { spanIsSampled } from '../utils/spanUtils';
 
 export const INTEGRATION_NAME = 'SpanStreaming' as const;
 
-export const spanStreamingIntegration = defineIntegration(() => {
+interface SpanStreamingOptions {
+  /**
+   * When enabled, a trace is flushed shortly after its segment span ends, rather than relying solely
+   * on the buffer's timeout/size thresholds or an explicit `flushTraceSpans` emission.
+   *
+   * This is used in the browser, where we never know when the page is closed or navigated away from,
+   * so spans must be sent timely.
+   *
+   * @default false
+   */
+  flushOnSegmentEnd?: boolean;
+}
+
+export const spanStreamingIntegration = defineIntegration((options: SpanStreamingOptions = {}) => {
   return {
     name: INTEGRATION_NAME,
 
@@ -22,6 +35,9 @@ export const spanStreamingIntegration = defineIntegration(() => {
       const buffer = new SpanBuffer(client);
 
       client.on('afterSpanEnd', span => {
+        // Negatively sampled spans must not be captured.
+        // This happens because OTel and we create non-recording spans for negatively sampled spans
+        // that go through the same life cycle as recording spans.
         if (!spanIsSampled(span)) {
           return;
         }
@@ -33,6 +49,18 @@ export const spanStreamingIntegration = defineIntegration(() => {
       client.on('flushTraceSpans', traceId => {
         buffer.flush(traceId);
       });
+
+      if (options.flushOnSegmentEnd) {
+        // In addition to capturing the span, we also flush the trace when the segment
+        // span ends to ensure things are sent timely. We never know when the browser
+        // is closed, users navigate away, etc.
+        client.on('afterSegmentSpanEnd', segmentSpan => {
+          const traceId = segmentSpan.spanContext().traceId;
+          setTimeout(() => {
+            buffer.flush(traceId);
+          }, 500);
+        });
+      }
     },
   };
 }) satisfies IntegrationFn;
