+++ packages/core/src/server-exports.ts
@@ -5,7 +5,6 @@
  */
 
 export { startSpan, startInactiveSpan, startSpanManual } from './tracing/trace';
-export { spanStreamingIntegration } from './integrations/spanStreaming';
 
 export type { ServerRuntimeClientOptions } from './server-runtime-client';
 export { ServerRuntimeClient } from './server-runtime-client';
